requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
appdirs==1.4.4
apturl==0.5.2
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
Authlib==0.15.5
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.8
checkbox-support==0.22
clang==5.0
click==7.1.2
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
common-utils-data==0.0.3
configparser==5.1.0
corider==0.1.5
crayons==0.4.0
cryptography==3.4.8
cycler==0.11.0
Cython==0.29.24
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distro-info===0.14ubuntu0.2
dnspython==2.1.0
docutils==0.18.1
entrypoints==0.3
et-xmlfile==1.1.0
exchangelib==3.3.2
feedparser==5.1.3
ffmpeg==1.4
flatbuffers==1.12
fonttools==4.28.2
fpdf2==2.4.6
frozenlist==1.2.0
fsspec==2021.11.1
furl==2.1.3
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
graphviz==0.13.2
grpcio==1.42.0
guacamole==0.9.2
h11==0.12.0
h5py==3.1.0
html5lib==0.999
httplib2==0.9.1
humanfriendly==10.0
idna==3.3
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.8.2
imutils==0.5.4
ipython-genutils==0.2.0
isodate==0.6.0
Jinja2==2.8
joblib==1.1.0
jsonpath-ng==1.5.3
jsonschema==4.1.0
jupyter-core==4.8.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
language-selector==0.1
louis==2.6.4
lxml==4.6.4
Mako==1.0.3
Markdown==3.3.6
MarkupSafe==0.23
mask-rcnn==2.1
matplotlib==3.5.0
mss==6.1.0
multidict==5.2.0
nest-asyncio==1.5.1
netsuitesdk==1.24.0
networkx==2.6.3
notifiers==1.2.1
ntlm-auth==1.5.0
numpy==1.21.4
oauthlib==3.1.1
onboard==1.2.0
opencv-python==4.5.3.56
openpyxl==3.0.9
opt-einsum==3.3.0
orderedmultidict==1.0.1
outcome==1.1.0
packaging==21.3
padme==1.1.1
pandas==1.3.4
parso==0.8.2
pdfminer.six==20201018
pexpect==4.0.1
pickleshare==0.7.5
Pillow==8.4.0
plainbox==0.25
platformdirs==2.4.0
ply==3.11
prompt-toolkit==3.0.20
protobuf==3.19.1
psutil==5.8.0
ptflops==0.6.6
ptyprocess==0.5
py-cpuinfo==8.0.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pycups==1.9.73
pycurl==7.43.0
pyDeprecate==0.3.1
Pygments==2.10.0
PyGObject==3.20.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyJWT==1.3.0
pynput-robocorp-fork==2.0.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PyPDF2==1.26.0
pyperclip==1.8.2
pyrsistent==0.18.0
pyserial==3.5
PySocks==1.7.1
python-apt==1.1.0b1+ubuntu0.16.4.12
python-dateutil==2.8.2
python-debian==0.1.27
python-dotenv==0.19.2
python-systemd==231
python-xlib==0.31
pytorch-lightning==1.4.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==6.0
pyzmq==22.3.0
reportlab==3.3.0
requests==2.26.0
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3987==1.3.8
ride==0.5.1
robotframework==4.1.2
robotframework-pythonlibcore==2.2.1
robotframework-requests==0.7.2
robotframework-seleniumlibrary==5.1.3
robotframework-seleniumtestability==1.1.0
rpa==1.45.0
rpaframework==9.5.0
rpaframework-core==6.5.1
rpaframework-pdf==0.5.1
rsa==4.8
scikit-image==0.18.3
scipy==1.7.3
seaborn==0.11.2
selenium==3.141.0
sessioninstaller==0.0.0
setuptools-scm==6.3.2
Shapely==1.7.1
simple-salesforce==1.11.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
ssh-import-id==5.5
supers==0.2.0
system-service==0.3
tagui==1.45.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tifffile==2021.10.12
tomli==1.2.2
torch==1.10.0
torchmetrics==0.6.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
trio==0.19.0
trio-websocket==0.9.2
tweepy==3.10.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.26.7
usb-creator==0.3.0
wcwidth==0.2.5
webdriver-manager==3.5.2
webdrivermanager==0.10.0
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.12.1
wsproto==1.0.0
wxPython==4.1.1
xdiagnose==3.8.4.1
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==0.7.3
xlutils==2.0.0
xlwt==1.3.0
yarl==1.7.2
zeep==4.1.0
zipp==3.6.0
